graph: convert legend.js to typescript
